APIs/requirements.txt
fastapi
uvicorn
pillow
opencv-python
numpy
torch
transformers
arabic-reshaper
python-bidi
pytesseract
paddleocr
paddlepaddle
ArabicOcr
langdetect
supervision
ultralytics
opencv-python-headless
pydantic
python-multipart
scikit-learn
